Move validate_error_trace_collector_json into validator directory
